requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.11.0
appnope==0.1.4
argcomplete==3.6.2
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.12.2
black==25.1.0
bleach==6.2.0
blinker==1.8.2
boto3==1.39.12
botocore==1.39.12
cachetools==6.2.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datamodel-code-generator==0.31.2
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.1.0
fastapi==0.121.2
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.7.0
fsspec==2024.10.0
gast==0.6.0
genson==1.3.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-genai==1.27.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.68.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.29.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
judgeval==0.1.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kagglehub==0.3.4
keras==3.7.0
kiwisolver==1.4.7
langchain==1.0.1
langchain-anthropic==0.3.17
langchain-core==1.0.0
langchain-huggingface==0.3.1
langchain-openai==0.3.28
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
litellm==1.74.8
llvmlite==0.44.0
lxml==6.0.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.97.1
opencv-python==4.10.0.84
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.0
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
pooch==1.8.2
praat-parselmouth==0.4.5
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.1
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.49.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.5.0
tiktoken==0.9.0
together==1.5.21
tokenizers==0.21.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typer==0.15.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.45.1
wrapt==1.17.0
xgboost==3.0.0
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
